feat: implementando action types
